norminette check

+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-d <lpaulo-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/29 14:46:13 by lpaulo-d          #+#    #+#             */
-/*   Updated: 2021/05/31 21:36:21 by lpaulo-d         ###   ########.fr       */
+/*   Updated: 2021/06/01 16:16:09 by lpaulo-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
+	size_t			i;
 	char			*ns;
 
 	i = 0;
